+++ lib/main.dart
@@ -1,16 +1,15 @@
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:look_back/settings/firebase_options.dart';
 import 'package:look_back/settings/theme_config.dart';
 import 'package:look_back/settings/model_theme.dart';
 import 'package:look_back/settings/routes.dart';
-import 'package:provider/provider.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'package:look_back/screens/messaging/firebase_options.dart';
 
 Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
   // If you're going to use other Firebase services in the background, such as Firestore,
   // make sure you call `initializeApp` before using other Firebase services.
-  
   print('WeeeeEEE');
   await Firebase.initializeApp();
   print("Handling a background message: ${message.messageId}");
@@ -76,7 +75,6 @@ class MyApp extends StatelessWidget {
           title: 'L O O K - B A C K',
           theme: getCurrentTheme(themeNotifier),
           routes: getApplicationRoutes(),
-
 
           // home: Center(
           //   child: Column(
