+++ lib/screens/signup/signup_form.dart
@@ -1,7 +1,10 @@
 import 'package:cool_alert/cool_alert.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:look_back/components/footer_sign.dart';
 import 'package:look_back/settings/theme_config.dart';
+import 'package:image_picker/image_picker.dart';
+import 'dart:io';
 
 class SignUpForm extends StatefulWidget {
   const SignUpForm({
@@ -175,7 +178,7 @@ class _SignUpFormState extends State<SignUpForm> {
           AlreadyHaveAnAccountCheck(
             login: false,
             press: () {
-              Navigator.pushNamed(context, '/login');
+              Navigator.pop(context);
             },
           ),
           const SizedBox(height: defaultPadding),
@@ -185,35 +188,108 @@ class _SignUpFormState extends State<SignUpForm> {
   }
 }
 
-class AlreadyHaveAnAccountCheck extends StatelessWidget {
-  final bool login;
-  final Function? press;
-  const AlreadyHaveAnAccountCheck({
+class SignUpScreenTopImage extends StatefulWidget {
+  const SignUpScreenTopImage({
     Key? key,
-    this.login = true,
-    required this.press,
   }) : super(key: key);
 
   @override
+  State<SignUpScreenTopImage> createState() => _SignUpScreenTopImageState();
+}
+
+class _SignUpScreenTopImageState extends State<SignUpScreenTopImage> {
+
+  File? _image;
+  final _picker = ImagePicker();
+
+  // Implementing the gallery
+  Future<void> _openImagePicker() async {
+    final XFile? pickedImage =
+        await _picker.pickImage(source: ImageSource.gallery);
+    if (pickedImage != null) {
+      setState(() {
+        _image = File(pickedImage.path);
+      });
+    }
+  }
+
+  // Implementing the camera
+  Future<void> _openCameraPicker() async {
+    final XFile? pickedImage =
+        await _picker.pickImage(source: ImageSource.camera);
+    if (pickedImage != null) {
+      setState(() {
+        _image = File(pickedImage.path);
+      });
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Text(
-          login ? "Don't have an Account ? " : "Already have an Account ? ",
-          //style: const TextStyle(color: kPrimaryColor),
-        ),
-        GestureDetector(
-          onTap: press as void Function()?,
-          child: Text(
-            login ? "Sign Up" : "Sign In",
-            style: const TextStyle(
-              //color: kPrimaryColor,
-              fontWeight: FontWeight.bold,
+    return Column(
+      children: [
+        const SizedBox(height: defaultPadding/2),
+        
+        SizedBox(height: defaultPadding),
+        Row(
+          children: [
+            const Spacer(),
+            Expanded(
+              flex: 8,
+              child: Column(
+                children: [
+
+                  const Text(
+                    "Sign Up",
+                    style: TextStyle(fontWeight: FontWeight.bold, fontSize: 25),
+                  ),
+
+                  const SizedBox(height: defaultPadding / 2),
+                  Container(
+                    alignment: Alignment.center,
+                    width: 200,
+                    height: 200,
+                    color: Colors.black26,
+                    child: _image != null
+                        ? Image.file(_image!, fit: BoxFit.cover,)
+                        : Image.asset('assets/customs/bob_cholo.png', fit: BoxFit.cover,)
+                        // : const Text(
+                        //   'Please select an image',
+                        //   style: TextStyle(color: Colors.black),
+                        // ),
+                  ),
+                  const SizedBox(height: defaultPadding / 2),
+
+                  Row(
+                    children: <Widget>[
+                      Expanded(
+                        child: ElevatedButton.icon(
+                          onPressed: _openImagePicker,
+                          icon: const Icon(Icons.add_photo_alternate_rounded),
+                          label: const Text('Select An Image from Gallery'),
+                        ),
+                      ),
+                      const SizedBox(width: defaultPadding / 2),
+                      Expanded(
+                        child: ElevatedButton.icon(
+                          onPressed: _openCameraPicker,
+                          icon: const Icon(Icons.add_a_photo_rounded),
+                          label: const Text('Select An Image from Camera'),
+                        ),
+                      ),
+                    ],
+                  )
+                  
+                ],
+              ),
+              
             ),
-          ),
-        )
+            const Spacer(),
+          ],
+        ),
+        SizedBox(height: defaultPadding),
       ],
     );
   }
 }
+
