+++ pubspec.yaml
@@ -31,7 +31,6 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
@@ -47,6 +46,8 @@ dependencies:
   google_sign_in: ^6.1.0
   github_sign_in: ^0.0.4
   firebase_messaging: ^14.6.1
+  cloud_firestore: ^4.7.1
+  firebase_storage: ^11.2.1
 
   lottie: ^2.2.0
   flutter_svg: ^2.0.2
@@ -112,6 +113,11 @@ flutter:
       fonts:
         - asset: assets/fonts/Decipher.ttf
         - asset: assets/fonts/DecipherSymbols.ttf
+    
+    - family: SivarPro
+      fonts:
+        - asset: assets/fonts/SivarPro.ttf
+
   #   - family: Schyler
   #     fonts:
   #       - asset: fonts/Schyler-Regular.ttf
